+++ old/pdfmerge.py
@@ -1,4 +1,4 @@
-#!/Users/pf183229/code/ppp/.venv/bin/python3
+#!/home/pf183229/code/PPP/.venv/bin/python3
 
 import sys
 import argparse