+++ old/printydump.py
@@ -1,4 +1,4 @@
-#!/Users/pf183229/code/ppp/.venv/bin/python3
+#!/home/pf183229/code/PPP/.venv/bin/python3
 
 #		        This program is part of
 #          Paul's Preponderating Prepresser v1.1